kernel: printf add more specifiers parsing

+++ moose/include/kmem.h
@@ -6,7 +6,7 @@
 #define memset __builtin_memset
 #define memmove __builtin_memmove
 
-static inline u32 strlen(const char *str) {
+static inline size_t strlen(const char *str) {
     const char *cur = str;
     while (*++cur);
     return cur - str;
